first draft metric collection

+++ Common/Workload/Metrics/Latency.cs
@@ -4,7 +4,7 @@
 	(
         int tid,
 		TransactionType type,
-		long period
+		int period
 	);
 }
